DPNG-10731 Adjust Application.update_manifest method
Updated the k8s service includes and methods, as they were creating circular import.

Modified the update_manifest method, so it now supports modification of json manifest.

This method will add any envs to existing ones, replaces services defined in manifest
with the ones provided by the user and also replaces the application name.

Please note, that "services" and "env" fields are not supported by the new tap engine
and also their structure is not defined.