DPNG-12530 Don't remove application that was already pushed
In some tests we verify that we cannot push an application with the same name
twice. During the second push, the tap cli command fails (it's expected).
During the error recovery, we remove the application with the same name.

The problem was that we remove the original application (we remove applications
by name) by name, not id. As a result we are left with no applications.

From now on we check the error message. If it says that the application already
exists, then we don't remove it, we only raise an exception.