DPNG-12804 Application should only use id
When cURLing for application, we received both classId and id. At first
they were diffirent, but now they are the same and in future classId
will be removed.

This commit removed the classId reading from the response and leaves a
simple id of the application intact. This should fix the get_application
test.