DPNG-12835 Application does not read existing manifest file
Application class didn't read the existing manifest file.

Removed the save_manifest method (as it did not update the existing manifest)

Modified the update_manifest method, it should be now usable instead of
save_manifest.

It looks for the manifest file, opens it, updates the information and
saves it in the same file.