DPNG-15104 Stop license checker from salting
Script work in the followin way:
If the first line is greater than 1, than omit that line (usually #!/usr/...)

Now everything but the second line has to be identical

The line check if the year has any '-', or ',' or spaces or numbers. If so
that is correct.

However user might put for example 2542-900 as year range and it will be correct.